move source files into src/ folder
