+++ src/StarbucksInput.coffee
