+++ src/StarbucksParser.coffee
