+++ src/StringInput.coffee
