+++ src/StringInput.js
