+++ src/starbucks.js
@@ -60,7 +60,7 @@ import {
 
 import {
   config
-} from './starbucks.config.js';
+} from '../starbucks.config.js';
 
 import {
   foundCmd,
