+++ src/starbucks_commands.js
@@ -8,7 +8,7 @@ import {
 
 import {
   config
-} from './starbucks.config.js';
+} from '../starbucks.config.js';
 
 import {
   error,
