+++ test/04markdown.test.js
@@ -6,7 +6,7 @@ import {
   markdownify,
   disableMarkdown,
   enableMarkdown
-} from '../markdownify.js';
+} from '../src/markdownify.js';
 
 import {
   AvaTester
