+++ test/07states.test.coffee
@@ -1,9 +1,9 @@
 # 07states.test.coffee
 
 import {say, undef, nonEmpty} from '@jdeighan/coffee-utils'
-import {StarbucksInput} from '../StarbucksInput.js'
-import {StarbucksOutput} from '../Output.js'
-import {foundCmd, finished} from '../starbucks_commands.js'
+import {StarbucksInput} from '../src/StarbucksInput.js'
+import {StarbucksOutput} from '../src/Output.js'
+import {foundCmd, finished} from '../src/starbucks_commands.js'
 import {AvaTester} from '@jdeighan/ava-tester'
 import {init} from './test_init.js'
