+++ test/test_init.coffee
@@ -1,11 +1,10 @@
 # test_init.coffee
 
 import {strict as assert} from 'assert'
-import test from 'ava'
 import {setUnitTesting} from '@jdeighan/coffee-utils'
-import {disableMarkdown} from '../markdownify.js'
-import {disableBrewing} from '../brewCoffee.js'
-import {disableSassify} from '../sassify.js'
+import {disableMarkdown} from '../src/markdownify.js'
+import {disableBrewing} from '../src/brewCoffee.js'
+import {disableSassify} from '../src/sassify.js'
 
 export init = () ->
 	setUnitTesting(true)
