+++ test/test_init.js
@@ -4,23 +4,21 @@ import {
   strict as assert
 } from 'assert';
 
-import test from 'ava';
-
 import {
   setUnitTesting
 } from '@jdeighan/coffee-utils';
 
 import {
   disableMarkdown
-} from '../markdownify.js';
+} from '../src/markdownify.js';
 
 import {
   disableBrewing
-} from '../brewCoffee.js';
+} from '../src/brewCoffee.js';
 
 import {
   disableSassify
-} from '../sassify.js';
+} from '../src/sassify.js';
 
 export var init = function() {
   setUnitTesting(true);
